strike

+++ src/main/resources/db/changelog/v-1.0/100-create-user-resource-material-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE user_access_resource_material (
     user_id bigint NOT NULL,
-    resource_id INT NOT NULL,
+    resource_id bigint NOT NULL,
     get_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP NOT NULL,
     CONSTRAINT PK_UserAccessResourceMaterial PRIMARY KEY (user_id, resource_id, get_date),
     CONSTRAINT FK_UserAccessResourceMaterial_User FOREIGN KEY (user_id) REFERENCES users(id),
